requirements.txt
scholarly==1.7.11
python-dotenv==1.0.1 
arxiv==2.1.0 